backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
drf-yasg==1.21.7
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
# googlemaps==4.10.0
twilio==8.10.0
celery==5.3.4
redis==5.0.1
dj-database-url
